requirements.txt
# Data collection
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.0

# YouTube API
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Video / audio processing
yt-dlp>=2024.12.13
scenedetect[opencv]>=0.6.7
opencv-python>=4.9.0
Pillow>=10.3.0
librosa>=0.10.2
soundfile>=0.12.1
numpy>=1.26.0,<2.2

# Scheduling / infra
schedule>=1.2.2

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tenacity>=8.3.0

# MLX local inference stack (Apple Silicon)
mlx>=0.18.0
mlx-lm>=0.19.0

# Note: whisper.cpp is compiled into ./vendor/whisper.cpp by run.sh (not a pip package)
# Note: Ollama is a separate system app: https://ollama.com
# Note: musicgen-mlx is optional — run.sh attempts `pip install musicgen-mlx` and continues on failure
